ardrone_server: squash refactor-structure feature
  bb2f0d7 Merge branch 'refactor-structure'
  9edb795 move also interfaces (2/2)
  ab8fe99 move also interfaces (1/2)
  e65fad0 move headers to include tree: now headers have package prefix
