+++ src/stable/components/ardrone_server/include/ardrone_server/interfaces/cmdveli.h
@@ -17,30 +17,23 @@
  *       Alberto Martín Florido <almartinflorido@gmail.com>	
  */
 
-#ifndef _DRONE_EXTRAI_H_
-#define _DRONE_EXTRAI_H_
+#ifndef _DRONE_CMDVELI_H_
+#define _DRONE_CMDVELI_H_
 
-#include "../teleop_twist.h"
-#include <jderobot/ardroneextra.h>
+#include "ardrone_server/teleop_twist.h"
+#include <jderobot/cmdvel.h>
 #include <Ice/Ice.h>
 
-namespace ardrone_extra
+namespace cmdvel
 {
-	class ExtraI: virtual public jderobot::ArDroneExtra
+	class CMDVelI: virtual public jderobot::CMDVel
 	{
 		public:
-			ExtraI();
-			virtual ~ExtraI();
-			virtual void land(Ice::Current const & c);
-			virtual void takeoff(Ice::Current const & c);
-			virtual void reset(Ice::Current const & c);
-			virtual void toggleCam(Ice::Current const & c);
-			virtual void recordOnUsb(bool  record, Ice::Current const & c);
-			virtual void ledAnimation(Ice::Int type, Ice::Float duration, Ice::Float freq, Ice::Current const & c);
-			virtual void flightAnimation(Ice::Int type, Ice::Float duration, Ice::Current const & c);
-			virtual void flatTrim(Ice::Current const & c);			
+			CMDVelI();
+			virtual ~CMDVelI();
+			virtual Ice::Int setCMDVelData(jderobot::CMDVelDataPtr const & data, Ice::Current const & c);
 		private:
-					
+
 	};
 }
 #endif
